Hint action hamel (#52)
* Made HintAction. Trying to figure out where SvcFacade is sendig the 
request to.

* Corrected some things to make the hints run correctly. Currently have 
the basics down. The hint action is currently piggybacking on the step
completion, but it shouldn't be a full step completion, so I'll have to
fix that.

* Hints are working properly on shift bits. Piggybacking on the step 
completion shouldn't be a problem. All other views will need to have 
their own hint function in the ShaTuTutor file like hintShiftBits.

---------

Co-authored-by: ChayaBasha <59859876+ChayaBasha@users.noreply.github.com>